feat: documentation can now describe partials
So far, only use of shortcodes was possible. However, it should also be
possible to use partials to leverage primer.css.

Note: it was not trivial to implement because, by construction, content
can not call partials directly, only shortcodes. So I had to "transform"
a call to a shortcode into an "equivalent" call to a partial.

Note 2: this transformation cannot describe all possible use cases.
I will likely need to add more in the future, in order to take into
consideration other possible constructs.